Separate requirements.txt export logic from graph construction
